camera more descriptive naming
