+++ voxels/src/client/camera/mod.rs
@@ -1,7 +1,12 @@
-mod cmp;
-mod sys;
-mod tag;
+mod s_uniform;
+mod s_resize;
+mod c_cam;
+mod t_cam;
 
-pub use cmp::*;
-pub use sys::*;
-pub use tag::*;
+pub use self::
+{
+    s_uniform::*,
+    s_resize::*,
+    c_cam::*,
+    t_cam::*,
+};
