+++ voxels/src/client/camera/s_resize.rs
@@ -1,0 +1,34 @@
+use ezgame::ecs::*;
+use ezgame::window;
+
+/// updates the camera matrix on window resize
+pub struct SCameraResize;
+
+impl System for SCameraResize
+{
+    const EVENT: Event = window::evt::RESIZED;
+    const ORDER: Order = ord::MID;
+
+    fn exe() -> SysFn
+    {
+        // begin...
+        sys("camera_resize_system")
+        // components...
+        .with_query(<Write<super::CCamera>>::query())
+        // resources...
+        .read_resource::<window::RWindow>()
+        // system
+        .build(|_, world, window, query|
+        {
+            let size = window
+                .as_ref()
+                .unwrap()
+                .inner_size();
+
+            for mut cam in query.iter_mut(world)
+            {
+                cam.resize(size.width as f32, size.height as f32);
+            }
+        })
+    }
+}
