+++ voxels/src/client/camera/s_uniform.rs
@@ -1,43 +1,10 @@
 use ezgame::ecs::*;
 use ezgame::gfx::*;
-use ezgame::window;
 use ezgame::time;
 use ezmath::*;
 
-/// updates the camera matrix on window resize
-pub struct SCameraResize;
-
 /// updates the camera view-proj uniform
 pub struct SCameraUniform;
-
-impl System for SCameraResize
-{
-    const EVENT: Event = window::evt::RESIZED;
-    const ORDER: Order = ord::MID;
-
-    fn exe() -> SysFn
-    {
-        // begin...
-        sys("camera_resize_system")
-        // components...
-        .with_query(<Write<super::CCamera>>::query())
-        // resources...
-        .read_resource::<window::RWindow>()
-        // system
-        .build(|_, world, window, query|
-        {
-            let size = window
-                .as_ref()
-                .unwrap()
-                .inner_size();
-
-            for mut cam in query.iter_mut(world)
-            {
-                cam.resize(size.width as f32, size.height as f32);
-            }
-        })
-    }
-}
 
 impl System for SCameraUniform
 {
